no Jira - clean up unneeded casts, unnecessary imports, compare logic tests between boolean and true/false, replace a loop with arraycopy
git-svn-id: https://svn.apache.org/repos/asf/uima/uimaj/trunk@1641248 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-core/src/main/java/org/apache/uima/internal/util/StringToIntMap.java
@@ -65,7 +65,7 @@ public class StringToIntMap {
    *         <code>key</code> is actually defined in the map.
    */
   public int get(String key) {
-    Integer i = (Integer) this.map.get(key);
+    Integer i = this.map.get(key);
     if (i == null) {
       return DEFAULT_VALUE;
     }
@@ -82,7 +82,7 @@ public class StringToIntMap {
    * @return The previous value of <code>key</code>, if it was set. <code>0</code> else.
    */
   public int put(String key, int value) {
-    Integer i = (Integer) this.map.get(key);
+    Integer i = this.map.get(key);
     int rc;
     if (i == null) {
       rc = 0;
